Implementing instructions and documented

+++ include/memory.h
@@ -11,12 +11,12 @@ public:
 
     ~memory() = default;
 
-    size_t write (uint16_t* buffer, size_t buffer_size, uint16_t start_position);
+    size_t write (uint8_t* buffer, size_t buffer_size, uint16_t start_position);
 
     uint16_t read (uint16_t position);
 
 private:
-    std::vector <uint16_t> _ram {};
+    std::vector <uint8_t> _ram {};
 };
